requirements.txt
PyQt5>=5.15.0
pdfplumber>=0.9.0
pdf2image>=1.16.0
Pillow>=9.0.0
PyPDF2>=3.0.0
networkx>=2.6  # Optionnel pour le graphique
matplotlib>=3.5.0  # Optionnel pour le graphique
openpyxl>=3.0.0  # Pour lire les fichiers Excel
odfpy>=1.4.0  # Pour lire les fichiers ODS
